refactor(apigw-sqs-msg-filtering-cdk): rename stack file to match naming convention
