+++ apigw-sqs-msg-filtering-cdk/cdk/test/cdk.test.ts
@@ -1,6 +1,6 @@
 import * as cdk from 'aws-cdk-lib';
 import { Template } from 'aws-cdk-lib/assertions';
-import { ApigwSqsMsgFilteringCdkStack } from '../lib/apigw-sqs-msg-filtering-cdk-stack';
+import { ApigwSqsMsgFilteringCdkStack } from '../lib/apigw-sqs-msg-filtering-stack';
 
 
 test('SQS Queue Created', () => {
